reduce usage of size_t

+++ hash.h
@@ -1,11 +1,11 @@
 #pragma once
 
-#include <stdlib.h>
+#include <stdint.h>
 
 // Fowler–Noll–Vo hash function
-static size_t strhash_fnv1a(const char *str)
+static uint64_t strhash_fnv1a(const char *str)
 {
-    size_t hash = 0xcbf29ce484222325;
+    uint64_t hash = 0xcbf29ce484222325;
     int c;
     while ((c = *str++)) {
         hash ^= c;
@@ -13,9 +13,9 @@ static size_t strhash_fnv1a(const char *str)
     }
     return hash;
 }
-static size_t memhash_fnv1a(const void *mem, size_t size)
+static uint64_t memhash_fnv1a(const void *mem, long size)
 {
-    size_t hash = 0xcbf29ce484222325;
+    uint64_t hash = 0xcbf29ce484222325;
     for (const char *byte = mem; byte < (char *)mem + size; ++byte) {
         hash ^= *byte;
         hash *= 0x00000100000001b3;
@@ -24,32 +24,32 @@ static size_t memhash_fnv1a(const void *mem, size_t size)
 }
 
 // Daniel J. Bernstein hash function
-static size_t strhash_djb2(const char *str)
+static uint64_t strhash_djb2(const char *str)
 {
-    size_t hash = 5381;
+    uint64_t hash = 5381;
     int c;
     while ((c = *str++)) hash = ((hash << 5) + hash) + c;
     return hash;
 }
-static size_t memhash_djb2(const void *mem, size_t size)
+static uint64_t memhash_djb2(const void *mem, long size)
 {
-    size_t hash = 5381;
+    uint64_t hash = 5381;
     for (const char *byte = mem; byte < (char *)mem + size; ++byte)
         hash = ((hash << 5) + hash) + *byte;
     return hash;
 }
 
 // hash function of sdbm database library, also used by gawk
-static size_t strhash_sdbm(const char *str)
+static uint64_t strhash_sdbm(const char *str)
 {
-    size_t hash = 0;
+    uint64_t hash = 0;
     int c;
     while ((c = *str++)) hash = c + (hash << 6) + (hash << 16) - hash;
     return hash;
 }
-static size_t memhash_sdbm(const void *mem, size_t size)
+static uint64_t memhash_sdbm(const void *mem, long size)
 {
-    size_t hash = 0;
+    uint64_t hash = 0;
     for (const char *byte = mem; byte < (char *)mem + size; ++byte)
         hash = *byte + (hash << 6) + (hash << 16) - hash;
     return hash;
